Created persistant state, added features to message sending
As the chat wouldn't persist a reload I used the persist functionality
from zustand/middleware to make the chatId state persistent.
Now the messages can be sent by just clicking enter after entering
message.
Image to be sent would be seen above the message box and can be sent
with or without messages.